requirements.txt
selenium
pytest
python-dotenv
pytest-html
pyclean